5/28 Update: Implemented kalman.c/h, kalmanV2.c/h
Implemented kalman.c/h directly from the original code, kalmanV2.c/h are for optimization.

Still have issues of redefining values.
