+++ alpaca-bms-FE9_Rev2/alpaca-bms-FE9_Rev2.cydsn/cyapicallbacks.h
@@ -16,10 +16,14 @@
 #define PCAN_RECEIVE_MSG_current_CALLBACK
 #define PCAN_RECEIVE_MSG_vehicle_state_CALLBACK
 #define PCAN_RECEIVE_MSG_charger_CALLBACK
+// Defined to enable interrupts for Kalman Filter
+#define KalmanFilt_Int_INTERRUPT_INTERRUPT_CALLBACK
     
 extern void PCAN_ReceiveMsg_current_Callback();
 extern void PCAN_ReceiveMsg_vehicle_state_Callback(); 
 extern void PCAN_ReceiveMsg_charger_Callback();
+// Defined to enable interrupts for Kalman Filter
+extern void KalmanFilt_Int_Interrupt_InterruptCallback();
 
 #endif /* CYAPICALLBACKS_H */   
 /* [] */
