+++ alpaca-bms-FE9_Rev2/alpaca-bms-FE9_Rev2.cydsn/kalmanV2.c
@@ -1,0 +1,35 @@
+/* ========================================
+ *
+ * Copyright YOUR COMPANY, THE YEAR
+ * All Rights Reserved
+ * UNPUBLISHED, LICENSED SOFTWARE.
+ *
+ * CONFIDENTIAL AND PROPRIETARY INFORMATION
+ * WHICH IS THE PROPERTY OF your company.
+ *
+ * ========================================
+*/
+
+/*
+#include "kalman.h"
+#include "project.h"
+
+void KalmanFilt_Int_Interrupt_InterruptCallback() {
+    CyGlobalIntDisable;
+    EKF();
+    // Setting the xhatk_1 and Pk_1 values
+	for (i = 0; i < 2; i++) {
+		// set values of xhatk_1 to xhatCorrected
+		xhatk_1[i][0] = xhatCorrected[i][0];
+		// For each column of second matrix
+		for (j = 0; j < 2; j++) {
+			// set values of Pk_1 to PCorrected
+			Pk_1[i][j] = PCorrected[i][j];
+		}
+	}
+    
+    CyGlobalIntEnable
+}
+*/
+
+/* [] END OF FILE */
